refactor(intelligence): add useNl2SqlChat engine, migrate widget (phase 2)
Add src/views/intelligence/useNl2SqlChat.js: the shared conversation engine
(provider/model config, topic list state + helpers, message hydration/loading,
and the send -> deliver -> stream -> reconcile -> detach/cancel lifecycle,
including the leave-while-running idiom for switch/new/delete).

WidgetChat.vue now consumes the engine and keeps only its widget-specific glue:
demo/mock send, analytics tracking, outbound + imperative controls, the
floating/inline shadow-DOM template, and the history-drawer close. Divergent
concerns are passed as options (getAgentId, emitEvent, notifyError,
reloadTopicsAfterRun, messagePageSize, topicTitleLength).

Adds useNl2SqlChat.spec.js (send / detach-on-new / cancel). All widget tests
stay green.
